main and gallery_detail modified

+++ WebContent/Gallery/gallery_detail.jsp
@@ -83,11 +83,20 @@
 			
 		})	
 		
+		
+		
+		
 		$("#delete").on("click",function(){
-			location.href="delete.gal?seq=${content.seq}"
 			
+			let result = confirm("정말로 삭제하시겠습니까?")
+            if(result){
+               location.href="delete.gal?seq=${content.seq}";
+            }else {
+            location.href="#";
+         }
 			
 		})
+		
 		$("#back").on("click",function(){
 			/* location.href="javascript:history.back()" */
 			location.href ="${pageContext.request.contextPath}/galList.gal?cpage=1"
@@ -212,7 +221,7 @@
 
 				<div class="row m-0">
 					<div class="col-12 col-lg-3 col-xl-2 p-0">
-						<a href="beforeLogin.gal?cpage=1"><img src="project_logo.jpg"></a>
+						<a href="forGallery.gal?cpage=1"><img src="project_logo.jpg"></a>
 					</div>
 					<div class="col-3 col-lg-2 col-xl-1 p-0 navitext">
 						<a href="${pageContext.request.contextPath}/getCourse.cos?course_area=종로구">산책장소</a>
@@ -308,22 +317,18 @@
 				<div class="col-lg-8 ftco-animate" style="margin: auto;">
 					<p>
 					<c:forEach var="img" items="${ilist }">
-						<img src="files/${img.oriName }" style="width:800px;height:700px; margin-left:-40px;">
+						<img src="files/${img.oriName }" style="width:800px;height:700px; margin-left:-30px;">
 					</c:forEach>
-						<!-- <img src="images/image_1.jpg" alt="" class="img-fluid"> -->
+						
 					</p>
 					<h2 class="mb-3">${content.title }</h2>
 
 					<p>${content.contents }</p>
 
 					<div class="about-author d-flex p-4 bg-light">
-						<div class="bio mr-5">
-							<img src="images/person_1.jpg" alt="Image placeholder"
-								class="img-fluid mb-4" style="width: 100px; height: 100px;'">
-						</div>
+						
 						<div class="desc">
-							<h3>${content.writer }</h3>
-
+							<h5>작성자 : ${content.writer }</h5>
 						</div>
 					</div>
 					<br><br>
@@ -337,6 +342,14 @@
 							<input type="button" value="목록으로" class="btn btn-primary"
 								id="back">
 
+						</c:when>
+						<c:when test="${login.id == 'admin' }">
+						
+						<input type="button" value="삭제" class="btn btn-primary"
+								id="delete">
+						<input type="button" value="목록으로" class="btn btn-primary"
+								id="back">
+						
 						</c:when>
 						<c:otherwise>
